Preliminary work on a new fast Blatter-Pattyn solver option
I created a new option, which_ho_approx = HO_APPROX_FAST_BP = 4.
I don't know yet if it will work.

The basic idea is to solve a vertically integrated 2D equation for
the velocity u_k at a single level, and then solve a tridiagonal equation
for delta_u = u - u_k in each column, ultimately converging on the
BP solution.  If it works, this could be much faster than the standard 3D
BP solve.

With this commit I implemented the new which_ho_approx option
and started filling in the required steps of the algorithm in
glissade_velo_higher.F90.  The code compiles but has not been tested.

There are no answer changes for standard cases (nor will there be
for subsequent work on this solver option).