Started implementing a Goldberg depth-integrated solver option
After the last commit, I debugged and tested the so-called 'fast BP'
option and found it not to work well.  I decided to switch to the
Goldberg (2011) depth-integrated solver option.  This is similar to
L1L2, but the 2D solve is for the depth-integrated mean velocity
instead of the basal velocity.

I wrote a coupled of new Goldberg subroutines and the code compiles,
but it has not yet been tested.