Debugging the Goldberg solver option
I got the Goldberg solver option working for some test problems.
Results for ISMIP-HOM Tests A and C are similar to published results
(Goldberg 2011) and generally are better than L1L2 results at the same
resolutions.

I verified that multiprocessor runs (on a Mac, up to 4 processors) give
the same results as single-processor runs, within machine roundoff, for
the dome problem.

There still seems to be a problem with the evolving dome.  After a few
timesteps, the 2D solver fails to converge in 100 iterations, and the
velocities are generally too slow compared to BP.  This needs to be studied further.