Changes needed for exact restart for Goldberg solver
The Goldberg solver needs the following fields for exact restart,
in addition to the standard Glissade restart fields:
uvel_2d, vvel_2d, btractx, btractx and efvs.  I added these fields
to the restart subroutine in glide_setup.F90.

We need uvel_2d and vvel_3d because these are typically the mean velocity
computed by the solver, and do not correspond to a particular level in the
3D velocity arrays.  The other fields are needed to initialize the vertical
shear, which is used to update the effective viscosity.

I added uvel_2d and vvel_2d to the velocity derived type and also to the list
of fields in glide_vars.def.  I assigned 'load = 1' to btractx and btracty to
make them loadable from the restart file.

I fixed a typo in the autogenerate utility.