Changed name of the Goldberg solver to depth-integrated viscosity approximation (DIVA)
The solver formerly known as the Goldberg approximation is now called
the depth-integrated viscosity approximation (DIVA).
I consulted with Dan, who requested that the name of the scheme should not
include his name.  "DIVA" seems like a reasonable acronym, since the
key features of the scheme are the integrals over viscosity (to obtain
the depth-integrated mean viscosity) and inverse viscosity (to obtain
the omega factors that relate the basal traction to the velocity
at different levels of the column).

I also changed 'efvs_integral' to 'omega' based on the notation in Goldberg (2011).
I noted in comments that what I call omega corresponds to Goldberg's omega/H.