Minor fix so that the DIVA solver works with a no-slip basal BC
Previously, the code was exiting with zero velocities when I tried to run DIVA
with a no-slip basal BC.  This was fixed with a minor change:

-       if (whichbabc == HO_BABC_NO_SLIP .and. solve_2d) then
+       if ((whichapprox==HO_APPROX_L1L2 .or. whichapprox==HO_APPROX_SSA) .and. &
+              whichbabc==HO_BABC_NO_SLIP) then

In other words, DIVA solves for the 2D velocity like L1L2 and SSA, but (unlike L1L2 and SSA)
still requires a 2D solve matrix solve when the basal velocity is zero.

Now the DIVA results with a no-slip basal BC (which_ho_babc = 6) are very similar
to results with a very large beta (which_ho_babc = 4), as expected.

I also did some minor cleanup of comments.