Changes so that DIVA works with nonzero Dirichlet BCs and velocity-dependent beta
Previously, uvel_2d and vvel_2d were not initialized correctly for the case of
prescribed nonzero velocity (e.g., the Ross and stream test cases).  Now, at
Dirichlet points, I set uvel/vvel_2d to the prescribed basal velocity in the
3D uvel/vvel arrays.  At least for now, I assume no vertical shear in Dirichlet
columns, so the basal velocity is the same as the velocity throughout the column.

Also, the basal velocity is needed as an input to calcbeta for certain choices
of which_ho_babc.  Previously, I passed in the 2D velocity solution, but for DIVA
this is the mean velocity and is not the correct input.  Now I declare
arrays ubas/vbas and fill them with the correct basal velocity for each
approximation (SSA, BP, DIVA, etc.).

For DIVA, this means that the basal velocity must now be computed after each
nonlinear iteration.  Before, it was computed only after the nonlinear solve had converged.

I verified that DIVA results are now very similar to BP results for both
the Ross and stream test cases.  The Ross case has prescribed nonzero velocities;
the stream case has both prescribed nonzero velocities and velocity-dependent beta.