Implemented a new parameterization for driving stress at the grounding line
For vertices adjacent to the grounding line, the surface elevation obtained
by taking differences in usrf between grounded cells and floating cells is often very large.
We can obtain a more realistic surface elevation gradient (and driving stress)
by evaluating the surface gradient (ds/dx)_g for grounded edges (i.e., edges
with grounded cells on each side) along with (ds/dx)_f for floating edges
(i.e., edges with floating cells on each side), and then taking ds/dx
at the vertex as the weighted average of (ds/dx)_g and (ds/dx)_f.
(And similarly for ds/dy.)
The weighting factor f_ground is computed in subroutine glissade_grounded_fraction,
which was added to the code a few months ago but until now hasn't been used.

With this commit I added a subroutine glissade_gradient_at_grounding_line in
module glissade_grid_operators.F90. This subroutine computes a corrected gradient
for vertices adjacent to the GL (i.e., vertices with 0 < f_ground < 1).
See the comments at the top of the subroutine for details about the algorithm.

This subroutine is called from subroutine glissade_velo_higher_solve, just after
computing the surface elevation gradient in the usual way.  The corrected surface
elevation gradients are then used to compute driving stresses.

Note: The GLP for driving stress will have no effect if which_ho_ground = 0,
which is the default value. It is activated by setting which_ho_ground = 1,
which computes values of f_ground such that 0 < f_ground < 1 at vertices
with a mix of grounded and floating neighbor cells. The gradient is corrected
only at vertices with 0 < f_ground < 1.

In subroutine glissade_grounded_fraction, I added f_pattyn = (rhow*b)/(rhoi*H) as
an output argument, because f_pattyn is needed as an input argument to
glissade_gradient_at_grounding_line. I changed 'fpat' to 'f_pattyn' throughout
the subroutine.

I tested the new grounding-line gradients for the MISMIP test case that Gunter
is developing. Gradients appear to be correct at the GL for the initial MISMIP profile, and
for some variations of this profile I implemented in order to test different parts
of the logic. For MISMIP geometry, it is expected that the GLP for driving stress
will lead to more realistic results at coarser resolutions.

I also tested the grounding-line gradients in the circular-shelf test case,
which has a grounding line near the center of the domain. For this problem the
answers are relatively insensitive to the value of which_ho_ground (0 or 1).

I modified a few Fortran formatting statement in glissade_transport.F90 and
glide_diagnostics.F90, to allow for smaller values of the diffusive CFL limit
and larger values of integer cell indices.

Also, I made a minor fix to the L1L2 solver so that the column velocities
are set to the basal velocity at vertices with a Dirichlet BC.  Previously, the code
was erroneously computing these velocities.