Modified driving and basal stresses for the grounding-line parameterization
I tested the driving stress GLP from the previous commit and found that it does not work.
It leads to large changes in ds/dx (and therefore driving stress) between one time step
and the next when a cell changes from floating to grounded or vice versa.
So I backed off to the previous method of estimating ds/dx for vertices adjacent to the GL
in the same way as for vertices far from the GL.  I convinced myself that a standard
centered difference gives the right answer if the goal is to compute the mean value
of ds/dx in the region between a grounded cell and a floating cell.

Also, I activated the GLP for basal stress by passing optional argument f_ground (the fractional grounded
area in the box surrounding a vertex) into subroutine basal_sliding_bc in glissade_velo_higher_solve.
At each vertex, beta is now multiplied by f_ground before doing the standard
finite-element integrals. When the GLP is turned off, f_ground = 1 for grounded cells and 0 for floating
cells, so there is no change in the answer.

I verified that results are BFB for the evolving dome test.

For MISMIP, I checked that results look qualitatively correct.  Specifically, I ran
the MISMIP experiment using the DIVA option (which allows dt = 0.2 yr at 1.6 km resolution)
and verified that (1) for standard parameter settings, the GL stabilizes not far from its input
geometry, and (2) under conditions that promote GL advance (smaller A or larger accumulation),
the GL advances smoothy.  More MISMIP testing is still to come.

Some specific changes:

* glide_types.F90: Added the flotation function f_pattyn to the geometry derived type.
This can be a useful diagnostic.

* glide_vars.def: Added f_ground and f_pattyn as output diagnostics

* glissade_masks.F90: Minor reorganization and cleanup of the f_pattyn calculation

* glissade_grid_operators.F90: Removed subroutine glissade_gradient_at_grounding_lines
after testing it and finding it not to work. Changed the name of subroutine
'glissade_edge_gradient' to 'glissade_gradient_at edges'.  Changed the dimensions
of edge gradients to (nx-1,ny) for df/ds and (nx,ny-1) for df/dy; these are the
dimensions of the edge gradients that can be correctly computed for a scalar field
of dimension (nx,ny).  Changed variable 'f' to 'field'.d

* glissade_velo_higher.F90: Added verbose_glp diagnostic option.  Added f_pattyn as a pointer.
Passed f_ground to basal_sliding_bc subroutine.  Added some verbose_beta diagnostics for the GLP.
Changed dimensions of edge gradients.

* glissade_velo_sia.F90: Changed dimensions of edge gradients.