Added a MISMIP test case directory
Gunter's branch (gunter1/mismip) already includes a mismip test directory.
I am creating a new one in this branch (lipscomb/glissade_velo) so as to start
with the minimum python scripts and config files needed to set up and
run a MISMIP test case. Other files will be added later, including one or
more scripts for continuing a MISMIP test suite that has already started.

The README file contains directions for running the Python script, mismipInit.py.

The Python script creates a .nc input file as required to run the test case.

The file mismipInit.config contains config options that result in what appears to
be stable and qualitatively correct behavior. Note in particular:

* I set ewn = 1321 and nsn = 5, with dew = dns = 1.6 km.  CISM requires at least 5 cells
in the vertical to enforce periodic NS boundary conditions.  The value of ewn will
vary with dew/dns.  Some commented-out values are included for 0.8 km and 3.2 km grids.

* I set idiag to correspond to a grid cell adjacent to the initial grounding line.

* I set which_ho_approx = 4, corresponding to the depth-integrated viscosity approximation
(DIVA) based on Goldberg (2011). For reasons not clear to me, DIVA is more stable than L1L2
and allows a time step of dt = 0.20 yr on the 1.6 km grid.  This is close to the advective
CFL limit.  DIVA is much faster than Blatter-Pattyn.  With these choices I can run for
several model decades in a few minutes on a Mac with 4 processors (and presumably much
faster on Mustang).

* The preconditioner is diagonal (which_ho_precond = 1).  It isn't very good for this problem,
but the velocity fields seem fairly realistic (though not fully converged).

* I set which_ho_ground = 1, which activates the new GLP and produces a more accurate basal
shear stress that varies smoothly in time, as compared to the case with no GLP.

* I set which_ho_assemble_beta = 1, so that the basal stress at each vertex is a function only
of beta at that vertex, instead of being smoothed as in a standard finite-element calculation.
This setting is better for resolving sharp changes in basal shear stress.

* I set glissade_maxiter = 50, not because this is sufficient to converge, but to save cycles.
The solver currently doesn't make much additional progress after 50 iterations.

* I set which_ho_resid = 4, since a relative convergence criterion makes more sense than an
absolute criterion for a problem to be run at several different resolutions.