Fix forcing to skip reading files with earliest time after current time
There is a bug in forcing where if the earliest time in the file is
after the current time, then the forcing routine will either read in the first
time level of the forcing file regardless of its value, or if the
forcing file is not the first forcing file specified in the config file
it will behave strangely and advance the time level read on every time
step.

This commit fixes those behaviors by only reading a forcing file if it
contains a valid time level, which is defined as a time level with a
time value before or equal to the current time.