Changes required for exact restart with MISMIP tests
Unlike most test problems (and realistic problems), the MISMIP domain has nonzero velocities
along the global boundary, with periodic BC at the north and south.
For this reason, exact restart requires that we write uvel and vvel to the restart file
on the extended staggered mesh, with an additional row of velocity points along the
north and east edges of the domain.

For DIVA, exact restart requires that in addition to the 3D uvel/vvel, we write the
uvel_2d and vvel_2d arrays, along with the basal traction arrays (btractx and btracty)
on the extended staggered mesh.

With this commit I've added the necessary code for exact restart using MISMIP or similar
problems with nonzero velocities along the boundary.  The way it works for the user is as
follows:

* When writing restart files, set restart_extend_velo in the 'options' section of the config file.
Then uvel_extend and vvel_extend are written to the restart file in lieu of uvel and vvel.
(We still write uvel and vvel to the standard history file; uvel_extend and vvel_extend are
written to history files only if specifically requested.)

* When reading a restart file containing uvel_extend and vvel_extend, there is nothing the
user needs to do.  These two fields are automatically read if present, and then (assuming
they contain at least one nonzero value) they are copied to uvel and vvel.

To make this work, I made the following changes:

* Added some *_extend fields in glide_vars.def.  (NOTE: A fresh configure will be needed.)

* Declared the new *_extend fields in glide_types.F90

* Added some logic in subroutine glide_define_restart variables to write the required
*_extend fields to the restart file.

* Added some logic in subroutine glissade_diagnostic_variable_solve to read in the *_extend
fields, if present, and do the appropriate copies and halo updates.

* Added some additional logic near the end of glissade_diagnostic_variable_solve to fill
the *_extend arrays as needed before output.

* Added some code in glide_types.F90 and glide_setup.F90 to support a new config option,
restart_extent_velo (= 1 for true, = 0 for false, with the default being false).

* Inserted comments to explain what's going on. I also fixed an erroneous comment in
subroutine calcbeta.

I verified that restart is now exact for MISMIP, using DIVA and SSA.

This work could have been avoided if uvel, vvel, etc. were declared with the same dimensions
as the scalar fields.  However, this design decision was made in early Glimmer days, and at
this point it would take a lot of work to change it.  (Now I sort of regret not doing that
work a couple of years ago.)