Improved the treatment of basal and driving stresses near the grounding line
Three major changes:
(1) New option for assembling the driving stress
(2) Modified computation of effective pressure in calcbwat
(3) Weighting of beta by f_ground in calcbeta

(1) In glide_types and glide_setup, I added a new option for computing the driving stress
in the Glissade velocity solver.  Previously, the driving stress was computed using a standard
finite-element approach, integrating over ds/dx in the elements adjacent to a given node.
This leads to significant smoothing, which is a problem near the GL because large values of ds/dx
just inland of the GL are smeared onto the shelf where the true value of ds/dx is much lower.
So I added an option for which only the local value of ds/dx contributes to the driving stress.

The old option is HO_ASSEMBLE_TAUD_STANDARD = 0, and the new option is HO_ASSEMBLE_TAUD_LOCAL = 1.
These are analogous to the two options for assembling basal stresses.  The old option remains
the default for now, but the new option works better for MISMIP (and likely for any problem with
an evolving GL). The new option is activated by setting which_ho_assemble_taud = 1 in the config file.

I added the driving-stress components taudx and taudy to glide_vars.def. (So a fresh configure is needed.)
These components are computed and scaled in subroutine glissade_velo_higher_solve.

(2) I modified the calculation of effective pressure N in calcbwat.  The biggest change was a bug fix.
Previously, model%geometry%thkmask was passed into subroutine glissade_stagger when interpolating
effecpress to effecpress_stag.  This is the wrong mask. Instead, we should pass in an ice mask whose value
is 1 where ice is present and 0 where ice is absent.

I also settled on a method for computing N for floating ice with p = 0. In this case the term (1 - Hf/H)^p = 0^0
is indeterminate. I opted for the convention N = rhoi*grav*H (i.e., 0^0 = 1) for all ice, even if floating.
This convention works well in combination with modification (3) below: multiplying beta by f_ground.
We have f_ground = 0 for fully floating ice, hence beta goes to 0 for floating ice as desired even
when N = rhoi*grav*H.  On the other hand, setting N = 0 beneath floating ice leads to large, sudden changes
in N when a cell switches from floating (Hf > H) to barely grounded (Hf slightly less than H).

I slightly reorganized the calculation to make more clear the capping of f_pattyn (i.e., Hf/H) between 0 and 1.

I also added halo updates for thck and topg that are required to compute N correctly in halo cells.

(3) I modified calcbeta to multiply beta by f_ground.  Previously, beta was set to zero beneath all floating ice,
with flotation defined by the Glide mask, GLIDE_IS_FLOAT(mask). This is not the correct approach with a GLP,
because the Glide mask does not resolve partial flotation fractions near the GL. With a GLP, a better approach
is to pass f_ground into calcbeta and multiply the computed value of beta by f_ground at each vertex. Then as
the GL advances and retreats, beta changes smoothly, as does the resulting basal shear stress.

I therefore added f_ground as an optional argument.  It is passed in from Glissade but not Glam.
With the GLP off, f_ground = 0 or 1 at all vertices, and the results are very similar to using the
Glide mask.  With the GLP on, beta transitions smoothly from its full value to 0 as we approach the GL from inland.

Previously, the GLP was implemented by multiplying beta by f_ground in subroutine basal_sliding_bc.
That multiplication is now redundant, so I removed it.  It is now assumed that the beta values passed into
the subroutine already have been weighted by f_ground.

Minor changes:
* I turned off some diagnostic output in the PCG subroutines. (These can be turned back on by setting verbose_pcg = T).
* I added some diagnostics for verbose_glp = T in glissade_velo_higher_solve.

I tested these changes in MISMIP with the following configurations, running anywhere from 3 to 30 model years:
* SSA, 1.6 km, p = 0
* SSA, 1.6 km, p = 1
* DIVA, 1.6 km, p = 0
* DIVA, 1.6 km, p = 1
* BP, 1.6 km, p = 0
* SSA, 3.2 km, p = 0
* SSA, 0.8 km, p = 0

In all cases the GLP was turned on, giving smooth variations in basal stress as the GL advances.
BP used the SIA-based preconditioner; other runs used the diagonal preconditioner.

The results are satisfactory in that the velocities and stresses evolve smoothly over time,
the GL advances steadily, and the velocity decreases from the initial large values as the surface
gradients near the GL are reduced.

At 1.6 km resolution, the SSA, DIVA and BP solvers are stable with a timestep of 0.25 yr.
Intuitively, I would expect the max allowed timestep to increase as dx increases.
However, the 3.2 km problem initially requires a smaller timestep (~0.125 yr) because of the steep geometry.
(The timestep probably could be increased once the initial gradients have smoothed out.)
The 0.8 km problem is unstable with a timestep of 0.25 yr but is stable with a timestep of 0.125 yr.

I have not yet run any cases for the hundreds to thousands of years required to reach steady state.
We should do this next to see how the results compare to Pattyn et al. (2012).