Modified the standard MISMIP config file based on recent changes and tests
I made the following changes in the default config file, mismipInit.config:

* Increased the timestep from dt = 0.20 yr to dt = 0.25 yr.
The model runs stably at 1.6 km for this slightly larger timestep using SSA, DIVA and BP.

* Changed the default approximation from DIVA to SSA.
Hopefully, the SSA will faithfully reproduce MISMIP results from Pattyn et al. (2012).
(This remains to be tested.) DIVA will not, because it includes vertical-shear terms
that are missing in the published MISMIP models.

* Set which_ho_assemble_taud = 1.  This is a new option which allows for local assembly
of the driving stress, without the undesirable smoothing in the standard FE approach.

* Added flwa_basal = 1.d-16.  This is the A_b term in the Schoof basal friction law.

* Changed the Coulomb bump wavelength to 2.0 m.  This is the standard value; the previous
value was compensating for a non-standard A_b term.

* Added some output variables that may be of interest (basal, driving and longitudinal stresses)

* Added an output file for restart, written every 1000 model years