Changed rhoi, rhoo, grav from parameters to variables
Until now, most physical constants used in CISM have been declared as parameters
in glimmer_physcon.F90.  These include rhoi = 910 kg/m3, rhoo = 1029 kg/m3,
and grav = 9.81 m/s2.

Some test problems, MISMIP in particular, specify different values. Until now, the
only way to change these values was to modify glimmer_physcon.F90 and recompile the code.
This can lead to confusion, because nothing in the model output states which values
of these constants were used for a particular run.

With this commit I changed rhoi, rhoo and grav to variables.  The default values are
set in glimmer_physcon.F90 and are the same as before.  These default values can now
be overridden by the user in the 'parameters' section of the config file.
E.g., for MISMIP the config file would have the following lines:
rhoi = 900.0
rhoo = 1000.0
grav = 9.80

The log file now reports the values of these variables.

Note that the defaults cannot be overridden in the config file if CISM is run
coupled to CESM, because CESM assigns values of rhoi, rhoo and grav in its shr_const_mod.

Given this change, I had to make small changes in several modules to remove parameters
that were dependent on rhoi and grav.  As a result, answers have changed at roundoff
level for Glide (in the evolving dome test).  Glissade answers remain BFB for the Halfar
and evolving dome tests.

There are also several Glam scaling parameters dependent on rhoi and grav.
To retain the same values for these scaling parameters, I defined them in terms
of new parameters rhoi_glam and grav_glam, whose values are the same as the CISM
default values of rhoi and grav.