Add glad_mbal_coupling, based on glint_mbal_coupling.
Note that this is greatly trimmed down, because it only contains stuff needed
for the SMB scheme.

Similarly, note that I have done away with the mbal instance that used to be a
member of glad_mbc, instead just inlining mbal%tstep, because that is the only
component of mbal needed here.