Added C_space_factor for MISMIP 3D experiments
The MISMIP 3D experiments include a 2D spatial factor (0 <= C_space_factor <= 1) that
multiplies the basal stress term C.  This factor is now an optional input field.
If it is present in the input file, it is read in on the unstaggered grid and
interpolated to the staggered grid at initialization.  If not present in the input file,
it is set to 1 everywhere, resulting in the same answers as would be obtained without
the space factor.

Based on work done initially by Gunter, I added C_space_factor in glide_vars.def.
It is declared as part of the basal_physics derived type in glide_types, and it is
added to the restart file for the relevant which_ho_basal BC options.
In subroutine glissade_initialise, this field gets a halo update and is interpolated
to the staggered_mesh.

I did a bit of cleanup along the way.  The glissade_stagger and glissade_unstagger subroutines
now default to stagger_margin = 0 instead of 1, which means that (as in the similar Glide
subroutine stagvarb) all values are used in the interpolation, and not just the values in cells
where ice is present.  If stagger_margin = 0, it is no longer necessary to pass in an ice mask
(which is only needed when stagger_margin = 1).

In glissade_initialise I replaced a call to the old stagvarb with the newer (but equivalent,
given stagger_margin = 0) glissade_stagger.

In glissade_therm.F90 I removed an unnecessary 'use' statement.

I verified that with these minor changes, results are BFB for the dome and circular_shelf tests.

I ran a MISMIP 3D standard experiment (C_space_factor = 1 everywhere) for 1000 years
and verified that there are no significant gradients in the y direction.  (The v velocities
are nonzero but are very small, ~10^(-9) m/yr.) Gunter was seeing gradients in earlier spinups,
which may have been related to the lack of a halo update for C_space_factor.

Because of the new input field in glide_vars.def, a fresh configure is needed.