Implemented a no-penetration global boundary condition
Until now, CISM has supported periodic global boundary conditions (by default),
as well as outflow boundary conditions (in which scalar values in the halo
are set to zero).

With this commit there is support for a no-penetration boundary condition,
GLOBAL_BC_NO_PENETRATION = 2.  The new option is enabled by setting global_bc = 2
in the config file.  For this condition the normal velocity is set
to zero at the global boundary of the domain.  That is, uvel = 0 at the east/west
boundaries, and vvel = 0 at the north/south boundaries.

This new BC was originally motivated by the MISMIP experiments, which have zero normal flux at the
north/south boundaries.  It is not strictly necessary to implement this BC, because vvel
at the boundary should be close to zero from the symmetry of the problem.
However, setting vvel = 0 can improve convergence and make it easier to obtain vvel ~ 0
throughout the domain.

I decided that the most robust way to implement the BC was to introduce two new masks,
umask_no_penetration and vmask_no_penetration, that are computed at initialization.
These masks have a value of 1 where the normal velocity is prescribed to be zero,
and elsewhere have a value of 0. This is the same numbering convention as kinbcmask.
However, setting kinbcmask = 1 implies both uvel = 0 and vvel = 0.  The no-penetration BC
implies either uvel = 0 or vvel = 0 at a given vertex on the global boundary, but not both.

Within glissade_velo_higher_solve, the two new masks are used together with kinbcmask
to implement Dirichlet BCs at selected vertices (for a 2D solve) or nodes (for a 3D solve).
Setting a single velocity component to zero, instead of both velocity components, requires
some new logic in subroutines dirichlet_boundary_conditions_2d and dirichlet_boundary_conditions_3d.

For now I have not changed the way halo quantities are computed for this BC.  Halo quantities
are still computed based on periodic BC.  For a symmetric problem like MISMIP or MISMIP3d, this results
in free slip tangential to the boundary.  For other problems, some more work may be needed to
ensure free slip and otherwise avoid difficulties with sharp scalar gradients at the global boundary.

I have not implemented the option of (say) no-penetration BC at north/south boundaries,
with outflow BC at east/west.  The MISMIP experiments work with no-penetration at east/west, provided the
calving parameters are set up to calve all ice before it reaches the east boundary.

Since the new logic is fairly intricate, I did not try to implement the new BC in glam.
It is supported only for glissade (with any approximation other than local SIA).

When using global_bc = 0 or 1, results are BFB. In particular, I verified that answers are unchanged
for the Ross test case (both 2D and 3D velocity solves).  The new Dirichlet logic reduces to the old logic
for the case that kinbcmask is applied and umask/vmask_no_penetration are not.

I tested the no-penetration option for the MISMIP and MISMIP3d test cases.
The main result is that fewer linear iterations to achieve convergence to a given accuracy.
For example, in the standard MISMIP test, ~1000 linear iterations are needed for the initial diagnostic
solve to drive vvel to a small value (small fractions of 1 m/yr).  With the default 200 iterations,
we see initial vvel ~10 m/yr (with variation in x but not y).  Using the no-penetration BC,
200 iterations are sufficient to drive vvel to ~0.01 m/yr for the initial solve and
lower values on subsequent solves.

I also did a quick test of the no-penetration BCs for a dome case with prescribed uniform eastward velocity
at all points except the east/west global boundaries.  Ice piles up at the east boundary, as expected.

I made the following changes in specific modules:

* glide_types.F90:
  - New global_bc option
  - New masks in velocity derived type

* glide_setup.F90:
  - Support for new BC.  CISM will abort if the new BC is requested for dycore
  options other than Glissade.

* parallel_mpi.F90 and parallel_slap.F90:
  - New subroutine, staggered_no_penetration_mask, to set up masks for the new BC
  - Also did some cleanup in subroutine distributed_grid

* glissade.F90:
  - Initialization for new BC option (call to staggered_no_penetration_mask)
  - Some cleanup of optional diagnostic prints

* glissade_velo.F90
  - Moved halo update for kinbcmask to glissade_initialise (does not change during runtime)

* glissade_velo_higher_pcg.F90:
  - Moved solver parameters (maxiters, tolerance, and solve_ncheck) from individual
  subroutines to the top of the module (above 'contains')

* glissade_velo_higher.F90
  - Added support for no_penetration masks.  These masks are used to construct umask_dirichlet
  and vmask_dirichlet (which are now integer instead of logical masks).
  - Modified the dirichlet subroutines to support the case that either uvel or vvel = 0
  (but not necessarily both) at a given node or vertex
  - Modified the L1L2 computation of 3D velocities for the same reason
  - Modified the verbose_dirichlet diagnostic prints
  - Added verbose_pcg option for diagnostic prints