Change intent(out) to intent(inout) for array outputs in glad_gcm.
I realized that these need to be inout so that they maintain their original
values when this routine is called but it's not time for an ice dynamics time
step. This happens when we're doing daily coupling (e.g., for an IG, FG or BG
compset).

The fact that this didn't cause problems before probably just implies that the
intel compiler keeps intent(out) arguments at their original values. But this
isn't necessarily the case for other compilers.

Note that these arguments had intent(inout) in glint. I probably mis-interpreted
the need for (inout) as opposed to (out) and so changed this initially. But now
I see why these need to be (inout), so I am restoring them to how they are in
glint.