Added a Glissade calving module and cleaned up the calving options
[Bill Sacks: Please see the note at the bottom of this commit message.]

I added a new module, glissade_calving.F90.  This module currently includes a single
subroutine, glissade_calve_ice, which reproduces the calving options from Glide and
adds logic for one new option:

  integer, parameter :: CALVING_NONE = 0
  integer, parameter :: CALVING_FLOAT_ZERO = 1
  integer, parameter :: CALVING_FLOAT_FRACTION = 2
  integer, parameter :: CALVING_RELX_THRESHOLD = 3
  integer, parameter :: CALVING_TOPG_THRESHOLD = 4
  integer, parameter :: CALVING_HUYBRECHTS = 5
  integer, parameter :: CALVING_DAMAGE = 6

Options 0-5 are already supported by Glide.  These are implemented almost identically for Glissade.
(The only difference is that Glissade masks are used in place of Glide masks to identify
calving cells.)  The calving criteria are local, so these options are trivial to parallelize.

Option 6 is a new option to be added by Jeremy Bassis and collaborators.
This option will use a scalar damage tracer.  Ice in a grid cell will calve (I think) provided that
(1) It is sufficiently damaged (however the damage threshold is defined), and
(2) It either borders the ocean, or is connected to the ocean by a path of sufficiently
    damaged grid cells.

These options used to be called MARINE_NONE, etc.  I changed 'MARINE' to 'CALVING'.
However, I did not change the 'marine_limit' or marine_margin' labels in the config files,
since this would involve changing a number of standard config options.
(We can decide later whether to do this.)

I also cleaned up module glide_ground.F90.  (At some point we might change the module and file names.)
I changed subroutine glide_marinlim to glide_calve_ice, and removed some unused subroutines
and in/out arguments.
This subroutine is largely redundant with glissade_calving, but I'm keeping it for now so that
previous Glide options are supported and the Glissade calving options can be revised and
extended independently.

In glide_types I added a new glide_calving type.

The calving type holds two fields: both defined on the ice grid:
(1) calving_thck = the thickness of ice that calves (typically the full ice thickness,
    but can be less for the CALVING_FLOAT_FRACTION option, and possibly for a future
    subgrid-scale calving parameterization)
(2) damage = a scalar describing the damage in a particular cell.
    The damage field will later be prognosed based on the dynamics (e.g., strain rates)
    and transported with other tracers.  For now it is simply initialized to zero.
The 'calving_thck' field was previously called 'calving' and was part of the climate derived type.
The damage field is new.  It is included in glide_vars.def.

The calving type also holds three parameters, all of which can be specified in
the 'parameters' section of the config file:
(1) calving_fraction = fractional thickness in floating cells that calves
(2) marine_limit = threshold in basal topography at which ice calves (formerly called mlimit)
(3) damage_threshold = value at which ice is deemed to be sufficiently damaged to calve
Previously, calving_fraction was the fractional thickness remaining after calving; now it denotes
the fractional thickness calved.
The first two parameters were previously part of the numerics derived type, and the third is new.

I tested calving options 0-5 for a few timesteps in a standard (1d) MISMIP test.
I verified that the new subroutine glissade_calve_ice gives the same results as the old subroutine
glide_calve_ice for each calving option.

NOTE TO BILL SACKS: I made a minor change in glint_upscale.F90 (changing calving to calving_thck).
Will we need to duplicate this change in Glad?