Added arrays for tracer values at upper and lower surfaces
In subroutine glissade_add_smb of module glissade_transport.F90,
any accumulation is added at the upper and lower ice surfaces.
This accumulation must be associated with a value for each tracer.

Previously, the temperature or enthalpy at each surface was passed into
the subroutine and used to update tracer #1 (which was hard-coded to be
temp/enthalpy).  Other tracer values were assumed to be zero at the surfaces.

Now, the tracer surface values are assigned in subroutine glissade_transport_setup_tracers.
They are put in generic arrays tracers_lsrf and tracers_usrf, so there is no
specific reference to the temp/enthalpy in the main transport driver.
This means that temperature does not have to be passed in explicitly.
The new arrays are part of the geometry derived type.

Based on Jeremy Bassis's recommendation, I specified that the damage at the upper surface
is equal to the damage in the top layer, and the damage at the lower surface is equal
to the damage in the lowest layer. In other words, if the ice is damaged, then new snow/ice
will not heal it, and if the ice is undamaged, new snow/ice will not increase the damage.
Other hypotheses may be tested later.

Note: There is a one-step time lag, in that the damage associated with new accumulation
is the old-time (pre-remap) damage value, not the new-time (post-remap) value.
I didn't see a clean way to use the new value. We can revisit this later if needed.

I verified that answers are BFB for a standard EISMINT-2 problem with surface accumulation
and temperature advection.