Added an ice age tracer
I added a tracer for ice age.  An ice age variable has been defined in the code for a
long time but has never been transported properly.  With the recent changes to support
a calving damage tracer, it was easy to add an age tracer.

The tracer is called 'ice_age' and is part of the geometry derived type.
It used to be called 'age', but 'ice_age' is easier to grep for.

There is a new option called which_ho_ice_age with two values:
  integer, parameter :: HO_ICE_AGE_NONE = 0
  integer, parameter :: HO_ICE_AGE_COMPUTE = 1
The default value is 1, meaning that ice age is computed.
But the computation can be turned off if ice age is not a desired diagnostic.
Note: ice age can be transported only with the glam and glissade dycores.

I added ice_age to glide_vars.def so that it can be read from and written to
history and restart files.  It is output in units of years.
If which_ho_ice_age = 1, then ice_age is an automatic restart variable.
It is diagnostic only, so omitting it from the restart file would not change the
prognostic ice state, but would reset the age to zero.

Ice age is incremented in glissade_tstep. For each column with nonzero thickness,
the ice age is incremented by dt with each time step. For columns with zero thickness,
ice age is set to zero.  This means that if ice reappears in a column from which it
has disappeared, the age starts over from zero.

If the age computation is on, then ice age is transported conservatively as a tracer,
along with temperature and other tracers. Accumulation at the upper and lower surfaces
arrives with age = 0.

I checked the calculation in a 20-ka EISMINT-2 simulation. The resulting age
field looks correct.  I verified exact restart; the ice age after 21 ka is
identical to the ice age after a 20-ka run plus a 1-ka restart.